WIP: automatically use Reqres where code is using it's own NSURLSession
This isn't complete, it's missing the bridging header etc. I use this
code in a project of my own, where Reqres wasn't able to log requests
from a third-party binary-only library. With this change, all those
requests are logged. If there's interest in merging this code I
could finish / clean it up.

Note that it uses ObjC to do the swizzling; it is possible to swizzle
in Swift but it fails into the problems mentioned in this article:

https://blog.newrelic.com/engineering/right-way-to-swizzle/

whether in practice that would be a problem in this case or not I don't
know, but when I'm doing slightly dodgy things I prefer to do it
the safest way possible.